requirements.txt
# requirements.txt
streamlit
sentence-transformers
faiss-cpu
PyMuPDF
openai
python-dotenv
numpy
torch